Wrapped Up Coiling Axis + Starting Coiling
Decided to implement the coiling axis by first specifying a normal axis
(in case it's not really normal it is projected onto a normal axis) and
then a binormal is calculated. The coiling rate is the degree the axis
rotates on each iteration, and the "normal" at any position is calculated
using the normal and binormal as generators and the coiling_rate * iteration
as the angle.